+++ examples/src/features/reentrancy_guard.rs
@@ -48,7 +48,7 @@ mod test {
 
     #[test]
     fn non_reentrant_function_can_be_called() {
-        let mut contract = ReentrancyMockDeployer::init(&odra::test_env());
+        let mut contract = ReentrancyMockDeployer::init(&odra_test::test_env());
         assert_eq!(contract.get_value(), 0);
         contract.non_reentrant_count();
         assert_eq!(contract.get_value(), 1);
@@ -56,7 +56,7 @@ mod test {
 
     #[test]
     fn ref_recursion_not_allowed() {
-        let mut contract = ReentrancyMockDeployer::init(&odra::test_env());
+        let mut contract = ReentrancyMockDeployer::init(&odra_test::test_env());
         assert_eq!(
             contract.try_count_ref_recursive(11).unwrap_err(),
             odra::ExecutionError::ReentrantCall.into()
@@ -65,7 +65,7 @@ mod test {
 
     #[test]
     fn local_recursion_allowed() {
-        let mut contract = ReentrancyMockDeployer::init(&odra::test_env());
+        let mut contract = ReentrancyMockDeployer::init(&odra_test::test_env());
         contract.count_local_recursive(11);
         assert_eq!(contract.get_value(), 11);
     }
