+++ modules/src/erc721_token.rs
@@ -161,7 +161,7 @@ mod tests {
     }
 
     fn setup() -> TokenEnv {
-        let env = odra::test_env();
+        let env = odra_test::test_env();
         TokenEnv {
             env: env.clone(),
             token: Erc721TokenDeployer::init(
